+++ api/api.js
@@ -2,9 +2,13 @@
 let MoSQL = require("../MoSQL.js");
 
 let API = (function(MoSQL){
-    let list = function(node,config) {
+    let model;
+    let setModel = (modelName) => {
+        model = MoSQL.createModel(modelName);
+    };
+    let list = function(node) {
         let connection = node.ConnectionService.getConnection(node.mysql);
-        let sql = config.sql.list;
+        let sql = `select * from ${model.viewName}`;
         let data = [];
 
         return connection.runSql(sql).then((response) => {
@@ -16,57 +20,56 @@ let API = (function(MoSQL){
         });
     };
 
-    let create = function(node,config) {
+    let create = function(node) {
         let sql = "";
-        let t = MoSQL.createModel(config.modelName);
         if (node.post){
             let connection = node.ConnectionService.getConnection(node.mysql);
-            t.setDBAll(node.post);
-            sql = t.toInsertSQL();
+            model.setDBAll(node.post);
+            sql = model.toInsertSQL();
             //console.log('insert task',sql);
-            let strName = config.recordName(node.post);
-            return connection.runSql(config.sql.exist.replace('{0}',node.post[config.pkField])).then((response) => {
+            let strName = model.recordName();
+            return connection.runSql(model.sqlSelect()).then((response) => {
                 if (response.err){
-                    //console.log(`got this error trying to validate if this ${config.recordRef} already exists`,err);
+                    //console.log(`got this error trying to validate if this ${strName} already exists`,err);
                     return false;
                 }
 
                 if (response.rows.length > 0){
-                    //console.log(`${config.recordRef} already exists: ${strName}`);
+                    //console.log(`${model.tableName} already exists: ${strName}`);
                     return false;
                 } else {
                     return response;
                 }
             }).then((response) => {
                 if (!response){
-                    return {operationOk: false, message: `${config.recordRef} already exists. id: ${strName}`};
+                    return {operationOk: false, message: `${model.tableName} already exists. id: ${strName}`};
                 }
 
                 return connection.runSql(sql,(responseCreate) => {
                     connection.close();
                     if (responseCreate.err){
-                        return {operationOk: false, message: `Error on ${config.recordRef} creation. id: ${strName}`};
+                        return {operationOk: false, message: `Error on ${model.tableName} creation. id: ${strName}`};
                     } else {
-                        return {operationOk: true, message: `${config.recordRef} created correctly. id: ${strName}`};
+                        return {operationOk: true, message: `${model.tableName} created correctly. id: ${strName}`};
                     }
                 });
             });
         }
     }
 
-    let update = function(node,config) {
-        let t = MoSQL.createModel(config.modelName);
-        let tWithChanges = MoSQL.createModel(config.modelName);
+    let update = function(node) {
+        let t = MoSQL.createModel(model.name);
+        let tWithChanges = MoSQL.createModel(model.name);
         if (node.post){
             let connection = node.ConnectionService.getConnection(node.mysql);
             
-            return connection.runSql(config.sql.exist.replace('{0}',node.post[config.pkField])).then((response) => {
-                let strName = config.recordName(node.post);
+            return connection.runSql(model.sqlSelect()).then((response) => {
+                let strName = model.recordName();
                 if (response.err){
                     console.log(`You try an update on a task that does not exist in the server. id: ${strName}`);
                     // create it
-                    create(node,config).then((responseCreate) => {
-                        let msg = `You try an update on a ${config.recordRef} that does not exist in the server. id: ${strName}. Tried to create it.`;
+                    create(node).then((responseCreate) => {
+                        let msg = `You try an update on a ${model.tableName} that does not exist in the server. id: ${strName}. Tried to create it.`;
                         return {operationOk: responseCreate.operationOk, message: `${msg} ${responseCreate.message}`};
                     });
                 }
@@ -79,23 +82,24 @@ let API = (function(MoSQL){
                 connection.runSql(sql).then((responseUpdate) => {
                     connection.close();
                     if (responseUpdate.err){
-                        return {operationOk: false, message: `Error on ${config.recordRef} modification. id: ${strName}`};
+                        return {operationOk: false, message: `Error on ${model.tableName} modification. id: ${strName}`};
                     } else {
-                        return {operationOk: true, message: `${config.recordRef} updated correctly. id: ${strName}`};
+                        return {operationOk: true, message: `${model.tableName} updated correctly. id: ${strName}`};
                     }
                 });
             });
         }
     }
 
-    let api = function(method,node,config) {
-        return this[method](node,config).then(data => {
+    let api = function(method,node) {
+        return this[method](node).then(data => {
             node.response.end(JSON.stringify(data));
         });
     };
 
     return {
-        list
+        setModel
+        , list
         , create
         , update
         , api