+++ api/movementAPI.js
@@ -2,34 +2,23 @@
 let MoSQL = require("../MoSQL.js");
 let baseAPI = require("./api.js");
 
-let API = (function(MoSQL,baseAPI){
-    let config = {
-        tableName: 'vimovement'
-        , modelName: 'Movement'
-        , recordName: (r) => `${r.mov_id} / ${r.mov_date} / ${r.mov_txt_account} / ${r.mov_txt_type} / ${r.mov_amount}`
-        , pkField: 'mov_id'
-        , recordRef: 'Movement'
-        , sql: {
-            list: `select * from vimovement`
-            , exist: `select * from vimovement where mov_id = '{0}'`
-        }
-    };
+let API = (function(MoSQL, baseAPI){
+	baseAPI.setModel('Movement');
+	let list = function(node) {
+		baseAPI.api('list', node);
+	}
 
-    let list = function(node) {
-        baseAPI.api('list',node,config);
-    };
+	let create = function(node) {
+		baseAPI.api('create', node);
+	}
 
-    let create = function(node) {
-        baseAPI.api('create',node,config);
-    };
+	let update = function(node) {
+		baseAPI.api('update', node);
+	}
 
-    let update = function(node) {
-        baseAPI.api('update',node,config);
-    };
-
-    let batch = function(node) {
+	let batch = function(node) {
         let sql = "";
-        let model = MoSQL.createModel(config.modelName);
+        //let model = MoSQL.createModel(config.modelName);
         let connection, insertionsOk = 0, insertionsError = 0;
         let promiseQueue = [];
 
@@ -37,8 +26,8 @@ let API = (function(MoSQL,baseAPI){
             connection = node.ConnectionService.getConnection(node.mysql);
         }
         node.post.forEach((p) => {
-            if (p[config.pkField] !== ''){
-                model = MoSQL.createModel(config.modelName);
+            if (p[model.getPK()[0]] !== ''){
+                model = MoSQL.createModel(model.name);
                 model.setDBAll(p);
                 sql = model.toInsertSQL();
                 console.log('insert model',sql);
@@ -63,11 +52,11 @@ let API = (function(MoSQL,baseAPI){
         });
     };
 
-    return {
-        list
-        , create
-        , update
-        , batch
-    };
-})(MoSQL,baseAPI);
+	return {
+		list
+		, create
+		, update
+		, batch
+	}
+})(MoSQL, baseAPI);
 module.exports = API;